requirements.txt
altgraph @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
amqp==5.2.0
arabic-reshaper==2.1.1
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
certifi==2020.12.5
chardet==4.0.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distlib==0.3.8
dj-database-url==0.5.0
Django==2.1.5
django-autoslug==1.9.8
django-autoslug-iplweb==1.9.4
django-bootstrap4==0.0.7
django-ckeditor==6.0.0
django-cors-headers==2.2.0
django-crispy-forms==1.10.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-widget-tweaks==1.4.8
djangorestframework==3.8.2
djangorestframework-simplejwt==5.3.1
filelock==3.14.0
future @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
gunicorn==20.0.4
html5lib==1.1
idna==2.10
jet==1.0
kombu==5.3.7
macholib @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
more-itertools==10.2.0
packaging==24.0
Pillow==8.2.0
platformdirs==4.2.2
pluggy==0.13.1
prompt-toolkit==3.0.43
psycopg2==2.8.6
py==1.11.0
PyJWT==1.7.1
PyPDF2==1.26.0
pytest==5.4.1
pytest-django==3.9.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-decouple==3.0
python-dotenv==0.12.0
pytz==2020.4
razorpay==1.2.0
reportlab==3.5.56
requests==2.25.1
six==1.14.0
sqlparse==0.4.1
static3==0.7.0
twilio==6.51.1
typing_extensions==4.11.0
tzdata==2024.1
Unipath==1.0
urllib3==1.26.2
vine==5.1.0
virtualenv==20.26.2
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==5.0.1
xhtml2pdf==0.2.5
